+++ packages/uui-file-dropzone/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './uui-file-dropzone.element';