+++ packages/uui-input-file/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './uui-input-file.element';