+++ packages/uui-input/lib/UUIInputEvent.ts
@@ -3,4 +3,5 @@ import { UUIInputElement } from './uui-input.element';
 
 export class UUIInputEvent extends UUIEvent<{}, UUIInputElement> {
   public static readonly CHANGE: string = 'change';
+  public static readonly INPUT: string = 'input';
 }